Enhance knowledge base creation with PQ compression

+++ src/redi/knowledge_base.py
@@ -6,6 +6,7 @@ import numpy as np
 import json
 from redi.utils import load_jsonl
 from tqdm import tqdm
+import math
 
 def create_kb(
         key_margin_steps: int = 10,
@@ -17,9 +18,13 @@ def create_kb(
         prompts_from_path: str = "prompts.jsonl",
         prompts_to_path: str = "new_prompts.jsonl",
 
-        key_compression: str | None = None,
+        use_pca: bool = False,
         embedding_dim: int = 512,
-        compression_path: str = "compression.pkl"
+        pca_path: str = "compression.pkl",
+        
+        use_pq: bool = False,
+        n_pq_centroids: int = 256,
+        n_pq_subvectors: int = 16,
 ) -> None:  
     """
     Create a knowledge base from a trajectory file and a prompt file.
@@ -32,17 +37,15 @@ def create_kb(
         faiss_index_path (str): The path to the faiss index file.
         prompts_from_path (str): The path to the prompts file.
         prompts_to_path (str): The path to the new prompts file.
-        key_compression (str | None): The type of compression to use for the keys. 
-                                      If None, no compression is used. 
-                                      If "pca", PCA compression is used.
+        use_pca (bool): Whether to use PCA for compression.
         embedding_dim (int): The dimension of the embedding space.
-        compression_path (str): The path to save the PCA compression model.
+        pca_path (str): The path to save the PCA compression model.
         
     Returns:
         None
     
     """
-    if key_compression is None:
+    if not use_pca:
         embedding_dim = 1 * 4 * 64 * 64
 
     prompts_from = load_jsonl(prompts_from_path)
@@ -51,28 +54,48 @@ def create_kb(
     with h5py.File(trajectory_path, "r") as trajectory_file:
         traj_names = list(trajectory_file.keys())
 
-        all_keys = []
 
         # PCA keys compression
-        if key_compression == "pca":
+        if use_pca:
+            all_keys = []
             pca = PCA(n_components=embedding_dim, svd_solver="auto")
             for name in traj_names:
                 flat_key = trajectory_file[name][key_margin_steps].reshape(1, -1)
                 all_keys.append(flat_key)
 
             all_keys = np.vstack(all_keys)
+            print("Fitting PCA...")
             pca.fit(all_keys)
-            with open(compression_path, "wb") as pca_file:
+            with open(pca_path, "wb") as pca_file:
                 pickle.dump(pca, pca_file)
-
-        index = faiss.IndexFlatL2(embedding_dim)
+        
+        # PQ keys compression
+        if use_pq:
+            n_bits = math.log2(n_pq_centroids)
+            if n_bits % 1 != 0:
+                raise ValueError("n_pq_centroids must be a power of 2.")
+            index = faiss.IndexPQ(embedding_dim, n_pq_subvectors, int(n_bits))
+            
+            all_keys = []
+            for name in traj_names:
+                flat_key = trajectory_file[name][key_margin_steps].reshape(1, -1)
+                if use_pca:
+                    flat_key = pca.transform(flat_key)
+                all_keys.append(flat_key)
+                
+            all_keys = np.vstack(all_keys)
+            print("Training PQ index...")
+            index.train(all_keys)
+                
+            
+        else:
+            index = faiss.IndexFlatL2(embedding_dim)
         index = faiss.IndexIDMap(index)
 
         with h5py.File(kb_path, "a") as kb_file:
             for id, name in tqdm(enumerate(traj_names)):
                 flat_key = trajectory_file[name][key_margin_steps].reshape(1, -1)
-
-                if key_compression == "pca":
+                if use_pca:
                     flat_key = pca.transform(flat_key)
 
                 index.add_with_ids(flat_key, np.array([id], dtype=np.int64))
@@ -91,7 +114,7 @@ def clean_kb(
         kb_path: str = "knowledge_base.h5",
         faiss_index_path: str = "faiss_index.bin",
         prompts_path: str = "prompts.jsonl",
-        compression_path: str = "compression.pkl",
+        compression_path: str | None = "compression.pkl",
 ) -> None:
     """
     Clean the knowledge base by removing the trajectory file and the faiss index file.
@@ -109,5 +132,6 @@ def clean_kb(
     os.remove(kb_path)
     os.remove(faiss_index_path)
     os.remove(prompts_path)
-    os.remove(compression_path)
+    if compression_path is not None:
+        os.remove(compression_path)
     print(f"Removed {kb_path}, {faiss_index_path}, {prompts_path}, {compression_path}")
